rename name of linter rule
