+++ azurerm/internal/services/iothub/iothub_resource.go
@@ -263,7 +263,7 @@ func resourceArmIotHub() *schema.Resource {
 							Optional:         true,
 							DiffSuppressFunc: suppressIfTypeIsNot("AzureIotHub.StorageContainer"),
 						},
-						//lintignore:AZURERMS001
+						//lintignore:AZRMS001
 						"encoding": {
 							Type:     schema.TypeString,
 							Optional: true,
