+++ azurermproviderlint/passes/AZRMR001/AZRMR001_test.go
@@ -1,0 +1,18 @@
+package AZRMR001_test
+
+import (
+	"testing"
+
+	"github.com/terraform-providers/terraform-provider-azurerm/azurermproviderlint/passes/AZRMR001"
+	"golang.org/x/tools/go/analysis/analysistest"
+)
+
+func TestAZRMR001(t *testing.T) {
+	testdata := analysistest.TestData()
+	analysistest.Run(t, testdata, AZRMR001.Analyzer, "a")
+}
+
+func TestAZRMR001Fixes(t *testing.T) {
+	testdata := analysistest.TestData()
+	analysistest.RunWithSuggestedFixes(t, testdata, AZRMR001.Analyzer, "a")
+}
