+++ azurermproviderlint/passes/AZRMR001/testdata/src/a/vendor
