+++ azurermproviderlint/passes/AZRMS001/testdata/src/a/vendor
