+++ azurermproviderlint/passes/checks.go
@@ -1,12 +1,12 @@
 package passes
 
 import (
-	"github.com/terraform-providers/terraform-provider-azurerm/azurermproviderlint/passes/AZURERMR001"
-	"github.com/terraform-providers/terraform-provider-azurerm/azurermproviderlint/passes/AZURERMS001"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurermproviderlint/passes/AZRMR001"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurermproviderlint/passes/AZRMS001"
 	"golang.org/x/tools/go/analysis"
 )
 
 var AllChecks = []*analysis.Analyzer{
-	AZURERMR001.Analyzer,
-	AZURERMS001.Analyzer,
+	AZRMR001.Analyzer,
+	AZRMS001.Analyzer,
 }
